marketplace/requirement.txt
django
djangorestframework
django-filter
cryptography
psycopg2
uwsgi
Pillow
